+++ l10n_bo_hr_payroll/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from . import hr_employee
 from . import hr_payroll
