+++ l10n_bo_hr_payroll/models/hr_payroll.py
@@ -38,22 +38,3 @@ class HrPayrollClosingTable(models.Model):
                     raise ValidationError(_('No puedes crear dos registro para el trabajador en el mismo periodo de tiempo.'))
 
 
-class HrPayrollQuinquennialData(models.Model):
-    _name = 'hr.payroll.quinquennial.data'
-    _description = 'Datos Quinquenio'
-    _rec_name = "employee_id"
-
-    payslip_id = fields.Many2one('hr.payslip')
-    contract_id = fields.Many2one('hr.contract', 'Contract', required=True)
-    employee_id = fields.Many2one('hr.employee', 'Employee', required=True)
-    # Campo calculado entre la fecha de consulta y la fecha de inicio de contrato menos la cantidad
-    # de años pagados
-    Balance = fields.Float(string="Saldo", required=True)
-    type_of_payment = fields.Selection([
-        ('0001', 'Pago de Quinquenio'),
-        ('0002', 'Adelanto de Quinquenio')],
-        string='Clase de documento', default='0001')
-    years_of_service =  fields.Integer(string="Años de antiguedad", required=True)
-    date_from = fields.Date(string='Fecha de inicio', help="Fecha de inicio de formulario de pago de quinquenio", required=True)
-    date_to = fields.Date(string='Fecha fin', help="Fecha fin de formulario de pago de quinquenio", required=True)
-    date_pay = fields.Date(string='Fecha de pago',  help="Fecha de pago de quinquenio", required=True)
