+++ l10n_bo_hr_payroll/security/ir.model.access.csv
@@ -1,2 +1,3 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_hr_payroll_closing_table_manager,hr.payroll.closing.table.manager,model_hr_payroll_closing_table,hr_contract.group_hr_contract_manager,1,1,1,1
+access_hr_payroll_quinquennial_data_manager,hr.payroll.quinquennial.data.manager,model_hr_payroll_quinquennial_data,hr_contract.group_hr_contract_manager,1,1,1,1
