+++ l10n_bo_hr_payroll/views/hr_employee_view.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<odoo>
+
+ <!--         EMPLOYEE VIEW DATOS QUINQUENIO        -->
+    <record id="hr_employee_view_form_inherit" model="ir.ui.view">
+        <field name="name">hr.employee.view.form.inherit</field>
+        <field name="model">hr.employee</field>
+        <field name="inherit_id" ref="hr.view_employee_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//page[@name='hr_settings']" position="after">
+                <page name="hr_quinquenio" string="Datos quinquenio" groups="hr.group_hr_user">
+                    <group>
+                        <field name="balance"/>
+                        <field name="years_of_service"/>
+                        <field name="quinquennial_ids">
+                            <tree>
+                                <field name="contract_id"/>
+                                <field name="employee_id"/>
+                                <field name="type_of_payment"/>
+                                <field name="amount_years"/>
+                                <field name="date_from"/>
+                                <field name="date_to"/>
+                            </tree>
+                        </field>
+                    </group>
+                </page>
+            </xpath>
+        </field>
+    </record>
+</odoo>
